Nina/UserManagementv2/requirements.txt
django
djangorestframework
mysqlclient
drf-yasg